Refactor save system to allow different Save versions
