+++ oriedita/src/main/java/oriedita/editor/exception/NewerFileFormatException.java
@@ -1,0 +1,7 @@
+package oriedita.editor.exception;
+
+public class NewerFileFormatException extends FileReadingException {
+    public NewerFileFormatException(Throwable cause) {
+        super(cause);
+    }
+}
