+++ oriedita/src/main/java/oriedita/editor/save/SaveV2.java
@@ -1,0 +1,5 @@
+package oriedita.editor.save;
+
+public class SaveV2 extends SaveV1{
+
+}
